==> 2_data_explore.Rmd <==
---
title: "2_data_explore"
output: html_document
date: "2023-06-21"
---

BEFORE:

1_data_process.R

input data (.csv): @data_sdg381, @data_sdg381_rmnch, @data_sdg382

AFTER: none

# 1. SETUP

```{r message=FALSE, warning=FALSE}
# load packages
library(tidyverse)   #process dataframe

# load functon
source("./ref/Zuur et al., 2010/HighstatLib.r") 

# load data
dir_data <- "./data"

sdg381 <- read_csv(file.path(dir_data, "@data_sdg381.csv"))
sdg381_rmnch <- read_csv(file.path(dir_data, "@data_sdg381_rmnch.csv"))
sdg382 <- read_csv(file.path(dir_data, "@data_sdg382.csv"))

```

# 2. SDG 3.8.1

histogram

```{r}
## histogram
ggplot(data = sdg381, aes(x = sdg381)) + geom_histogram(bins = 30)
ggplot(data = sdg381, aes(x = log(sdg381))) + geom_histogram(bins = 30)

```

relationship - control variables

```{r}
ggplot(data = sdg381, aes(x = log(health.worker), y = log(sdg381))) + geom_point() 
ggplot(data = sdg381, aes(x = log(sdi), y = log(sdg381))) + geom_point() 
ggplot(data = sdg381, aes(x = log(government.health.spending), y = log(sdg381))) + geom_point() 
ggplot(data = sdg381, aes(x = log(prepaid.private.health.spending), y = log(sdg381))) + geom_point() 
ggplot(data = sdg381, aes(x = log(out.of.pocket.health.spending), y = log(sdg381))) + geom_point() 
```

relationship explanatory variables

```{r}
### total.aid
ggplot(data = sdg381, aes(x = log(total.aid), y = log(sdg381))) + geom_point() 
ggplot(data = sdg381, aes(x = log(total.aid.volatility), y = log(sdg381))) + geom_point() 
### total.aid.for.health
ggplot(data = sdg381, aes(x = log(total.aid.for.health), y = log(sdg381))) + geom_point() 
ggplot(data = sdg381, aes(x = log(total.aid.for.health.volatility), y = log(sdg381))) + geom_point() 
### total.aid.for.health
ggplot(data = sdg381, aes(x = institutional.index, y = log(sdg381))) + geom_point() 

```

# 3. SDG 3.8.1 - RMNCH

histogram

```{r}
## histogram
ggplot(data = sdg381_rmnch, aes(x = sdg381_rmnch)) + geom_histogram(bins = 30)
ggplot(data = sdg381_rmnch, aes(x = log(sdg381_rmnch))) + geom_histogram(bins = 30)

```

relationship - control variables

```{r}
ggplot(data = sdg381_rmnch, aes(x = log(health.worker), y = log(sdg381_rmnch))) + geom_point() 
ggplot(data = sdg381_rmnch, aes(x = log(sdi), y = log(sdg381_rmnch))) + geom_point() 
ggplot(data = sdg381_rmnch, aes(x = log(government.health.spending), y = log(sdg381_rmnch))) + geom_point() 
ggplot(data = sdg381_rmnch, aes(x = log(prepaid.private.health.spending), y = log(sdg381_rmnch))) + geom_point() 
ggplot(data = sdg381_rmnch, aes(x = log(out.of.pocket.health.spending), y = log(sdg381_rmnch))) + geom_point() 
```

relationship explanatory variables

```{r}
### total.aid
ggplot(data = sdg381_rmnch, aes(x = log(total.aid), y = log(sdg381_rmnch))) + geom_point() 
ggplot(data = sdg381_rmnch, aes(x = log(total.aid.volatility), y = log(sdg381_rmnch))) + geom_point() 
### total.aid.for.health
ggplot(data = sdg381_rmnch, aes(x = log(total.aid.for.health), y = log(sdg381_rmnch))) + geom_point() 
ggplot(data = sdg381_rmnch, aes(x = log(total.aid.for.health.volatility), y = log(sdg381_rmnch))) + geom_point() 
### total.aid.for.health
ggplot(data = sdg381_rmnch, aes(x = institutional.index, y = log(sdg381_rmnch))) + geom_point() 

```

# 4. SDG 3.8.2

histogram

```{r}
## histogram
ggplot(data = sdg382, aes(x = sdg382)) + geom_histogram(bins = 30)
ggplot(data = sdg382, aes(x = log(sdg382))) + geom_histogram(bins = 30)

```

relationship - control variables

```{r}
ggplot(data = sdg382, aes(x = log(health.worker), y = log(sdg382))) + geom_point() 
ggplot(data = sdg382, aes(x = log(sdi), y = log(sdg382))) + geom_point() 
ggplot(data = sdg382, aes(x = log(government.health.spending), y = log(sdg382))) + geom_point() 
ggplot(data = sdg382, aes(x = log(prepaid.private.health.spending), y = log(sdg382))) + geom_point() 
ggplot(data = sdg382, aes(x = log(out.of.pocket.health.spending), y = log(sdg382))) + geom_point() 
```

relationship explanatory variables

```{r}
### total.aid
ggplot(data = sdg382, aes(x = log(total.aid), y = log(sdg382))) + geom_point() 
ggplot(data = sdg382, aes(x = log(total.aid.volatility), y = log(sdg382))) + geom_point() 
### total.aid.for.health
ggplot(data = sdg382, aes(x = log(total.aid.for.health), y = log(sdg382))) + geom_point() 
ggplot(data = sdg382, aes(x = log(total.aid.for.health.volatility), y = log(sdg382))) + geom_point() 
### total.aid.for.health
ggplot(data = sdg382, aes(x = institutional.index, y = log(sdg382))) + geom_point() 

```

# 5. COLLINEARITY

## total.aid

```{r}
# total.aid
Z <- sdg381 %>% select(total.aid, total.aid.volatility,
                       institutional.index, gdp.per.capita, government.health.spending, 
                       prepaid.private.health.spending, out.of.pocket.health.spending, 
                       health.worker, sdi)
corvif(Z)

# remove health.worker
Z <- sdg381 %>% select(total.aid, total.aid.volatility,
                       institutional.index, gdp.per.capita, government.health.spending, 
                       prepaid.private.health.spending, out.of.pocket.health.spending, 
                       sdi)
corvif(Z)
```

## total.aid.for.health

```{r}
# total.aid.for.health
Z <- sdg381 %>% select(total.aid.for.health, total.aid.for.health.volatility,
                       institutional.index, gdp.per.capita, government.health.spending, 
                       prepaid.private.health.spending, out.of.pocket.health.spending, 
                       health.worker, sdi)
corvif(Z)

# remove health.worker
Z <- sdg381 %>% select(total.aid.for.health, total.aid.for.health.volatility,
                       institutional.index, gdp.per.capita, government.health.spending, 
                       prepaid.private.health.spending, out.of.pocket.health.spending, 
                       sdi)
corvif(Z)
```

## total.aid.for.health_rmnch

```{r}
# total.aid.for.health rmnch
Z <- sdg381 %>% select(total.aid.for.health_rmnch, total.aid.for.health.volatility_rmnch,
                       institutional.index, gdp.per.capita, government.health.spending, 
                       prepaid.private.health.spending, out.of.pocket.health.spending, 
                       health.worker, sdi)
corvif(Z)

# remove health worker
Z <- sdg381 %>% select(total.aid.for.health_rmnch, total.aid.for.health.volatility_rmnch,
                       institutional.index, gdp.per.capita, government.health.spending, 
                       prepaid.private.health.spending, out.of.pocket.health.spending, 
                       sdi)
corvif(Z)
```



```
